+++ demo/26-paragraph-borders.ts
@@ -1,7 +1,7 @@
 // Creates two paragraphs, one with a border and one without
 // Import from 'docx' rather than '../build' if you install from npm
 import * as fs from "fs";
-import { Document, Packer, Paragraph } from "../build";
+import { BorderStyle, Document, Packer, Paragraph } from "../build";
 
 const doc = new Document({
     sections: [
@@ -14,13 +14,13 @@ const doc = new Document({
                         top: {
                             color: "auto",
                             space: 1,
-                            value: "single",
+                            style: BorderStyle.SINGLE,
                             size: 6,
                         },
                         bottom: {
                             color: "auto",
                             space: 1,
-                            value: "single",
+                            style: BorderStyle.SINGLE,
                             size: 6,
                         },
                     },
