+++ demo/61-text-frame.ts
@@ -1,7 +1,16 @@
 // Text Frame (Text Box) example
 // Import from 'docx' rather than '../build' if you install from npm
 import * as fs from "fs";
-import { Document, FrameAnchorType, HorizontalPositionAlign, Packer, Paragraph, TextRun, VerticalPositionAlign } from "../build";
+import {
+    BorderStyle,
+    Document,
+    FrameAnchorType,
+    HorizontalPositionAlign,
+    Packer,
+    Paragraph,
+    TextRun,
+    VerticalPositionAlign,
+} from "../build";
 
 const doc = new Document({
     sections: [
@@ -29,25 +38,25 @@ const doc = new Document({
                         top: {
                             color: "auto",
                             space: 1,
-                            value: "single",
+                            style: BorderStyle.SINGLE,
                             size: 6,
                         },
                         bottom: {
                             color: "auto",
                             space: 1,
-                            value: "single",
+                            style: BorderStyle.SINGLE,
                             size: 6,
                         },
                         left: {
                             color: "auto",
                             space: 1,
-                            value: "single",
+                            style: BorderStyle.SINGLE,
                             size: 6,
                         },
                         right: {
                             color: "auto",
                             space: 1,
-                            value: "single",
+                            style: BorderStyle.SINGLE,
                             size: 6,
                         },
                     },
