--- src/file/paragraph/formatting/border-attributes.ts
@@ -1,15 +1,0 @@
-import { XmlAttributeComponent } from "file/xml-components";
-
-export class BorderAttributes extends XmlAttributeComponent<{
-    readonly color: string;
-    readonly space: number;
-    readonly val: string;
-    readonly sz: number;
-}> {
-    protected readonly xmlKeys = {
-        val: "w:val",
-        color: "w:color",
-        space: "w:space",
-        sz: "w:sz",
-    };
-}
