--- src/file/styles/border/index.ts
@@ -1,1 +1,0 @@
-export * from "./border-style";
