+++ src/file/styles/styles.ts
@@ -3,7 +3,6 @@ import { BaseXmlComponent, ImportedXmlComponent, XmlComponent } from "file/xml-c
 import { StyleForCharacter, StyleForParagraph } from "./style";
 import { ICharacterStyleOptions } from "./style/character-style";
 import { IParagraphStyleOptions } from "./style/paragraph-style";
-export * from "./border";
 
 export interface IStylesOptions {
     readonly default?: IDefaultStylesOptions;
